+++ onnxruntime/core/providers/armnn/tensor/concat.cc
@@ -29,7 +29,7 @@ Status Concat<T>::Compute(OpKernelContext* ctx) const {
 
   // Hold pointers to the input tensors to be used in the PrepareForCompute() step
   std::vector<const Tensor*> input_tensors;
-  input_tensors.reserve(static_cast<uint64_t>(input_count));
+  input_tensors.reserve(input_count);
 
   LOGS_DEFAULT(VERBOSE) << "Concat ArmNN:";
   for (int i = 0; i < input_count; ++i) {
@@ -47,11 +47,11 @@ Status Concat<T>::Compute(OpKernelContext* ctx) const {
 
     // Calculate the size of the concatenated axis
     size_t concat_axis_size = 0;
-    for (uint64_t index = 0; index < input_count; index++) {
-      concat_axis_size += static_cast<uint64_t>(input_tensors[index]->Shape()[static_cast<size_t>(axis_)]);
+    for (int64_t index = 0; index < input_count; index++) {
+      concat_axis_size += input_tensors[index]->Shape()[static_cast<int>(axis_)];
     }
 
-    output_dims[static_cast<size_t>(axis_)] = static_cast<int64_t>(concat_axis_size);
+    output_dims[axis_] = concat_axis_size;
   } else { // 'Stack' mode
     // While stacking, the rank of the output is one more than the input rank(s).
     // Stacking may be thought of as adding an unit dimension (of value 1) in the input tensors,
